tests: Move into tests/ and split into fixture and reporting tests
