+++ tests/conftest.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+import pytest
+
+pytest_plugins = 'pytester'
