+++ tox.ini
@@ -6,4 +6,4 @@ deps =
     py==1.4.30
     pytest==2.8.3
 commands =
-    {envpython} -m py.test {posargs:test_pytest_catchlog.py}
+    {envpython} -m py.test {posargs:tests}
